# swagger.yaml
...
openapi: 3.0.0
servers:
    - description: Development
      url: http://localhost:4000/api/v1
info:
    description: This is a online book shop API
    version: "1.0.0"
    title: Book-shop-api
    contact:
        email: alhabib607273@gmail.com
    license:
        name: Apache 2.0
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
    /users/{id}:
        get:
            security:
                - bearerAuth: []
            tags: [User]
            description: Retrive a single a user
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "200":
                    description: Successfully retrived a user
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        $ref: "#/components/schemas/User"

                "404":
                    $ref: "#/components/responses/404"
                "401":
                    $ref: "#/components/responses/401"
                "500":
                    $ref: "#/components/responses/500"
        patch:
            security:
                - bearerAuth: []
            description: Update a user
            tags: [User]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            requestBody:
                content:
                    "application/json":
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    example: habib
                                role:
                                    type: string
                                    example: customer
                                account:
                                    type: string
                                    example: 13sdsd30dd6149
            responses:
                "200":
                    description: User update successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        $ref: "#/components/schemas/User"
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                format: url
                                                example: "/users/:id"
                "400":
                    $ref: "#/components/responses/400"
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
        delete:
            security:
                - bearerAuth: []
            description: delete a User
            tags: [User]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "204":
                    description: User deleted successfully
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
...
    /carts/{id}:
        get:
            security:
                - bearerAuth: []
            tags: [Cart]
            description: Retrive a single a cart
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "200":
                    description: Successfully retrived a cart
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        allOf:
                                            - type: object
                                              properties:
                                                  id:
                                                      type: string
                                                      example: 23sd4dfdf34
                                                  user:
                                                      type: string
                                                      example: 12ds2123sd
                                                  books:
                                                      type: array
                                                      items:
                                                          type: string
                                                          example: 2kd3dfs3dsd23

                                                  quantity:
                                                      type: integer
                                                      format: int32
                                                      example: 5
                                                  amount:
                                                      type: integer
                                                      format: int32
                                                      example: 12000
                                            - $ref: "#/components/schemas/Timestamp"
                                    links:
                                        type: object
                                        properties:
                                            user:
                                                type: string
                                                format: url
                                                example: "/users/:id"
                "404":
                    $ref: "#/components/responses/404"
                "401":
                    $ref: "#/components/responses/401"
                "500":
                    $ref: "#/components/responses/500"
        patch:
            security:
                - bearerAuth: []
            description: Update a cart
            tags: [Cart]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                type: string
                                example: 64f167d158
            responses:
                "200":
                    description: Cart updated successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            allOf:
                                                - type: object
                                                  properties:
                                                      id:
                                                          type: string

                                                          example: 2sdf3sdf42
                                                      books:
                                                          type: array
                                                          items:
                                                              type: object
                                                              properties:
                                                                  id:
                                                                      type: integer
                                                                      format: int32
                                                                      example: 2342
                                                                  title:
                                                                      type: string
                                                                      example: clean code
                                                      quantity:
                                                          type: integer
                                                          format: int32
                                                          example: 3
                                                      amount:
                                                          type: integer
                                                          format: int32
                                                          example: 3240

                                                - $ref: "#/components/schemas/Timestamp"
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                format: url
                                                example: "/cart/:id"

                "400":
                    $ref: "#/components/responses/400"
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
        delete:
            security:
                - bearerAuth: []
            description: Delete a cart
            tags: [Cart]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "204":
                    description: Cart deleted successfully
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
...
    /orders/{id}:
        get:
            security:
                - bearerAuth: []
            description: Retrive a single Order
            tags: [Order]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "200":
                    description: Single article view
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        $ref: "#/components/schemas/Order"
                                    links:
                                        type: object
                                        properties:
                                            cart:
                                                type: string
                                                format: url
                                                example: "/carts/:id"
                                            user:
                                                type: string
                                                format: url
                                                example: "/users/:id"
                "400":
                    $ref: "#/components/responses/400"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
        patch:
            security:
                - bearerAuth: []
            description: Upadate a order
            tags: [Order]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                cartId:
                                    type: string
                                    example: g2d34ddff2f3
                                shippingMethod:
                                    type: string
                                    enum: [standard, priority]
                                    default: standard
                                orderStatus:
                                    type: string
                                    enum: [pending, cancelled]
                                    default: "pending"
            responses:
                "200":
                    description: Order updated successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        allOf:
                                            - type: object
                                              properties:
                                                  id:
                                                      type: string
                                                      example: 1sd2d3fdf2
                                                  cart:
                                                      type: string
                                                      example: 1sd2d3fdf2
                                                  quantity:
                                                      type: integer
                                                      example: 4
                                                  amount:
                                                      type: integer
                                                      example: 12000
                                                  orderStatus:
                                                      type: string
                                                      enum:
                                                          [
                                                              pending,
                                                              shipped,
                                                              delivered,
                                                              cancelled,
                                                          ]
                                                      default: pending
                                                  shippingMethod:
                                                      type: string
                                                      enum: [standard, priority]
                                                      default: standard
                                            - $ref: "#/components/schemas/Timestamp"
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                format: url
                                                example: "/order/:id"
                                            user:
                                                type: string
                                                format: url
                                                example: "/user/:id"
                                            cart:
                                                type: string
                                                format: url
                                                example: "/cart/:id"
                "400":
                    $ref: "#/components/responses/400"
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
        delete:
            security:
                - bearerAuth: []
            description: delete a book
            tags: [Order]
            parameters:
                - $ref: "#/components/parameters/idPathParam"
            responses:
                "204":
                    description: Order deleted successfully
                "401":
                    $ref: "#/components/responses/401"
                "404":
                    $ref: "#/components/responses/404"
                "500":
                    $ref: "#/components/responses/500"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        User:
            allOf:
                - type: object
                  properties:
                      id:
                          type: string
                          example: 3d4dfs5asd
                      name:
                          type: string
                          example: al-habib
                      email:
                          type: string
                          format: email
                          example: habib@gmail.com
                      role:
                          type: string
                          example: customer
                      # user here condition for account number ...important

                - $ref: "#/components/schemas/Timestamp"
        Order:
            allOf:
                - type: object
                  properties:
                      id:
                          type: string
                          example: 3sd4sd2f3f5
                      user:
                          type: string
                          example: 2dsk3d76ds
                      cart:
                          type: string
                          example: 2d3d1d23
                      quantity:
                          type: integer
                          example: 4
                      amount:
                          type: integer
                          example: 12000
                      orderStatus:
                          type: string
                          default: pending
                      shippingMethod:
                          type: string

                - $ref: "#/components/schemas/Timestamp"
...
        Timestamp:
            type: object
            properties:
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
            required: [createdAt, updatedAt]
    parameters:
...
        idPathParam:
            in: path
            name: id
            required: true
            description: Item Id
            schema:
                type: string
                example: 1d2sf3sf12f31dflsd242df
    responses:
        "500":
            description: Internal Server Error
            content:
                "application/json":
                    schema:
                        type: object
                        properties:
                            code:
                                type: integer
                                format: int32
                                example: 500
                            error:
                                type: string
                                example: Internal Server Error
                            message:
                                type: string
                                example: We are sorry for the inconvenience. Please try again later
        "400":
            description: Bad Request
            content:
                "application/json":
                    schema:
                        type: object
                        properties:
                            code:
                                type: integer
                                format: int32
                                example: 400
                            error:
                                type: string
                                example: Bad Request
                            data:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        field:
                                            type: string
                                            example: email
                                        message:
                                            type: string
                                            example: Invalid email
                                        in:
                                            type: string
                                            enum: [query, body]
                                            example: body
        "401":
            description: Unauthorized
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            code:
                                type: integer
                                format: int32
                                example: 401
                            error:
                                type: string
                                example: Unauthorized
                            message:
                                type: string
                                example: You don't have the right permissions
        "404":
            description: Recource Not Found
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            code:
                                type: integer
                                format: int32
                                example: 404
                            error:
                                type: string
                                example: Not Found
                            message:
                                type: string
                                example: Requested resource not found